-Release 모드 빌드 시, 발생하는 경고문들 수정 (초기화, 자료형 이슈 등) -잘모르지만 씬을 Release할때 두번 시도하는 부분을 발견, 해당 부분을 지움. (이게 원인인지는 잘 모르겠습니다.)

+++ AvoidTheBoss/Player.cpp
@@ -36,7 +36,7 @@ CPlayer::~CPlayer()
 /*플레이어의 위치를 변경하는 함수이다. 플레이어의 위치는 기본적으로 사용자가 플레이어를 이동하기 위한 키보드를
 누를 때 변경된다. 플레이어의 이동 방향(dwDirection)에 따라 플레이어를 fDistance 만큼 이동한다.*/
 
-void CPlayer::Move(const int8& dwDirection, float fDistance)
+void CPlayer::Move(const int16& dwDirection, float fDistance)
 {
 	XMFLOAT3 xmf3Shift = XMFLOAT3(0, 0, 0);
 	if (LOBYTE(dwDirection))
